test/smoke: extract macros to own file
