+++ test/smoke/smoke_test.go
@@ -9,26 +9,6 @@ import (
 	sous "github.com/opentable/sous/lib"
 )
 
-func initProject(t *testing.T, client *sousProject, flags *sousFlags, transforms ...ManifestTransform) {
-	client.MustRun(t, "init", flags.SousInitFlags())
-	client.TransformManifest(t, flags.ManifestIDFlags(), transforms...)
-}
-
-// initBuild is a macro to fast-forward to tests to a point where we have
-// initialised the project, transformed the manifest, and performed a build.
-func initBuild(t *testing.T, client *sousProject, flags *sousFlags, transforms ...ManifestTransform) {
-	initProject(t, client, flags, transforms...)
-	client.MustRun(t, "build", flags.SourceIDFlags())
-}
-
-// initBuildDeploy is a macro for fast-forwarding tests to a point where we have
-// initialised the project as a kind project, built it with -tag tag and
-// deployed that tag successfully to cluster.
-func initBuildDeploy(t *testing.T, client *sousProject, flags *sousFlags, transforms ...ManifestTransform) {
-	initBuild(t, client, flags, transforms...)
-	client.MustRun(t, "deploy", flags.SousDeployFlags())
-}
-
 func TestSmoke(t *testing.T) {
 
 	m := newRunner(t, matrix())
